requirements.txt
python-binance==1.0.25
pandas==2.2.3